requirements.txt
django
djangorestframework
selenium==4.33.0
psycopg2-binary
jdatetime
phonenumbers
django-phonenumber-field
django-environ
fake-useragent
webdriver-manager
undetected-chromedriver
selenium-stealth
Pillow
numpy
scipy
pymongo
python-dotenv
django-redis
concurrent-log-handler  # log in os level, so fastapi and django logd doesnt curropt each other
selenium-wire==5.1.0
blinker==1.4
mapbox-vector-tile
motor